+++ src/annotator.ts
@@ -1,14 +1,9 @@
 import { Walker } from './plainer';
-import { transformValue } from './transformer';
-import { FlattenAnnotations } from 'serializer';
+import { transformValue, untransformValue } from './transformer';
+import { FlattenAnnotations, TypeAnnotation } from './serializer';
+import { stringifyPath, parsePath } from './pathstringifier';
+import { mapDeep } from './mapDeep';
 
-const escapeKey = (key: string) => key.replace(/\./g, '\\.');
-
-const pathToKey = (path: (string | number)[]): string =>
-  path
-    .map(String)
-    .map(escapeKey)
-    .join('.');
 export const makeAnnotator = () => {
   const annotations: FlattenAnnotations = {};
 
@@ -23,7 +18,7 @@ export const makeAnnotator = () => {
           annotations.values = {};
         }
 
-        annotations.values[pathToKey(path)] = transformed.type;
+        annotations.values[stringifyPath(path)] = transformed.type;
       }
       return transformed.value;
     } else {
@@ -32,4 +27,30 @@ export const makeAnnotator = () => {
   };
 
   return { annotations, annotator };
+};
+
+export const applyAnnotations = (
+  plain: any,
+  annotations: FlattenAnnotations
+): any => {
+  if (annotations.values) {
+    const annotationsWithPaths = Object.entries(annotations.values).map(
+      ([key, type]) => [parsePath(key), type] as [string[], TypeAnnotation]
+    );
+    const annotationsWithPathsLeavesToRoot = annotationsWithPaths.sort(
+      ([pathA], [pathB]) => pathB.length - pathA.length
+    );
+
+    for (const [path, type] of annotationsWithPathsLeavesToRoot) {
+      plain = mapDeep(plain, path, v =>
+        untransformValue(v, type as TypeAnnotation)
+      );
+    }
+  }
+
+  if (annotations.root) {
+    plain = untransformValue(plain, annotations.root);
+  }
+
+  return plain;
 };
